combine event_flag files
